backend/app/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
motor==3.3.2
pyjwt==2.8.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.0